% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_packa.R
\name{packa_f1}
\alias{packa_f1}
\title{f1 exposed by packa}
\usage{
packa_f1()
}
\value{
A String "Hello from packa_f1"
}
\description{
This is function \code{f1} exposed by \code{packa}
}
\examples{
packa_f1()
}

